test: fix async test for fetching image
